Implemented file_stream with unit tests.
file_strewam is a wrapper around stream_parser specialized at
parsing a file in a br archive. It does magic parsing, determines
the endianness of the file, parses file meta in header such as the
version number and header size.

Pushes a scope of offset base in stream_parser at construction
time and pops at destruction time. Seeks the stream position to
the end of file.