Added sequence data type: varint.
A varint represents a variable-length unsigned integer. Each read
is one byte. If the most significant bit of the byte is 1, then
the parser should continue to read one more byte. the less signi-
ficant bits are then ORed into the result. This works similarly
to MIDI's variable length quantity.