Initial implementation of sequence event parsing.
Defined a list of known sequence events, each as an abstract class
with argument getters.

Using metaprogramming techniques, actual implementations can read
the correct number of arguments and types from the stream just by
specifying the event argument types as template arguments. Events
that take other events(e.g. random, variable) are also describable
using the same templates.

Support for labels is minimal because that is only more important
when supporting editing. For now the P0 use case is being able to
play the music, so labels are not as important and have no public
APIs to access.

Iterating the events in their correct types is done using the
visitor pattern.