Implemented data::types::sized_ref.
The sized_ref type is a common type that has two fields: offset to
some data, the length of the data. Since coming up with variable
names are difficult, made this a type so we can have less variables!

The type also supports the stream operator, so we are able to reduce
two stream calls to just one, which makes the code cleaner.